Convert archive-payload-result to skill with session-first architecture
The archive skill now derives everything from the original Claude session
tarball instead of calling live APIs. Adds curl shim, fixes extract script
to handle persisted-output references and object-format fetch_payloads
responses, and consolidates env vars to single EVAL_ARCHIVES_DIR.
