+++ eval.yaml
@@ -9,9 +9,6 @@ execution:
   max_budget_usd: 25.0
   env:
     PATH: $PATH
-    EVAL_ARTIFACT_ARCHIVE: $EVAL_ARTIFACT_ARCHIVE
-    EVAL_CACHED_RESPONSES: $EVAL_CACHED_RESPONSES
-    EVAL_GH_CACHE: $EVAL_GH_CACHE
     EVAL_ARCHIVES_DIR: $EVAL_ARCHIVES_DIR
 
 runner:
